Create CV_MLmodel_plus
This is an overview how my program works based on the dataset I've provided and train it giving me 3 outputs and the accuracy with which it evaluate the process and also the components.
Outputs are:
1. A general front 
2. The analysis of the given data having multiple jobs and job title that is being evaluated and shown as the bar graph format
3. The GUI which ---> which asks the user to define the jobs and also select the job category . upon evaluating the csv file which we select , it gives the data of the candidate that have applied for the job in the respective fields
4.  data represents the classification report for the given code. 

Accuracy of the model on the training set: 0.99
Accuracy of the model on the test set: 0.99


The classification report provides detailed information about the performance of the model for each class. Here's what each column in the classification report represents:

precision: This column shows the precision score for each class. Precision represents the proportion of true positive predictions out of all positive predictions. In other words, it measures the accuracy of the positive predictions.
recall: This column shows the recall score for each class. Recall represents the proportion of true positive predictions out of all actual positive instances. It measures the ability of the model to correctly identify positive instances.
f1-score: This column shows the F1-score for each class. The F1-score is the harmonic mean of precision and recall, providing a balanced measure of the model's performance.
support: This column shows the number of instances belonging to each class in the test set.